pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "stockbot"
version = "0.1.0"
description = "Autonomous trading bot with ML signal ensemble and RL execution"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",

    # Data ingestion
    "polygon-api-client>=1.14.0",
    "alpaca-py>=0.26.0",
    "yfinance>=0.2.40",
    "aiohttp>=3.9.5",
    "httpx>=0.27.0",

    # ML & models
    "torch>=2.3.0",
    "transformers>=4.41.0",
    "datasets>=2.19.0",
    "stable-baselines3>=2.3.0",
    "gymnasium>=0.29.1",
    "lightgbm>=4.3.0",
    "shap>=0.45.0",
    "scikit-learn>=1.4.2",

    # Feature engineering
    "pandas>=2.2.2",
    "pandas-ta>=0.3.14b",
    "numpy>=1.26.4",
    "scipy>=1.13.0",
    "ta-lib>=0.4.28",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",

    # Scheduling & queuing
    "apscheduler>=3.10.4",
    "redis>=5.0.4",
    "celery>=5.4.0",

    # Monitoring
    "prometheus-client>=0.20.0",
    "structlog>=24.1.0",

    # Config & utils
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "tenacity>=8.3.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.5",
    "mypy>=1.10.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
